Added 0x16 format
