+++ catena4450m101_sensor/assets/provisioned.png
