+++ catena4460_aqi/git-repos.dat
@@ -1,0 +1,13 @@
+#
+# list of git repositories needed by this directory.
+# Use git-boot.sh to be fetch them to the Arduino libraries directory.
+#
+https://github.com/mcci-catena/Adafruit_FRAM_I2C.git
+https://github.com/mcci-catena/Catena-Arduino-Platform.git
+https://github.com/mcci-catena/arduino-lorawan.git
+https://github.com/mcci-catena/Catena-mcciadk.git
+https://github.com/mcci-catena/arduino-lmic.git
+https://github.com/mcci-catena/Adafruit_BME680.git
+https://github.com/mcci-catena/Adafruit_Sensor.git
+https://github.com/mcci-catena/RTCZero.git
+https://github.com/mcci-catena/BH1750.git
