+++ extra/catena-message-0x16-decoder-node-red.js
@@ -1,0 +1,109 @@
+// JavaScript source code
+// This Node-RED decoding function decodes the record sent by the Catena 4450
+// Water-Pressure & Water-Level Sensing application.
+
+var b = msg.payload;  // pick up data for convenience; just saves typing.
+
+// an empty table to which we'll add result fields:
+//
+// result.vBat: the battery voltage (if present)
+// result.vBus: the USB charger voltage (if provided)
+// result.boot: the system boot counter, modulo 256
+// result.t: temperature in degrees C
+// result.p: station pressure in hPa (millibars). Note that this is not
+//   adjusted for the height above sealevel so can't be directly compared
+//   to weather.gov "barometric pressure"
+// result.rh: relative humidity (in %)
+// result.lux: light level, in lux
+// result.wPressure: water pressure value in Kpa.
+// result.wLevel: Water Level value.
+var result = {};
+
+// check the message type byte
+if (b[0] != 0x16) {
+    // not one of ours: report an error, return without a value,
+    // so that Node-RED doesn't propagate the message any further.
+    node.error("not ours! " + b[0].toString());
+    return;
+}
+
+// i is used as the index into the message. Start with the flag byte.
+var i = 1;
+// fetch the bitmap.
+var flags = b[i++];
+
+if (flags & 0x1) {
+    // set vRaw to a uint16, and increment pointer
+    var vRaw = (b[i] << 8) + b[i + 1];
+    i += 2;
+    // interpret uint16 as an int16 instead.
+    if (vRaw & 0x8000)
+        vRaw += -0x10000;
+    // scale and save in result.
+    result.vBat = vRaw / 4096.0;
+}
+
+if (flags & 0x2) {
+    var vRaw = (b[i] << 8) + b[i + 1];
+    i += 2;
+    if (vRaw & 0x8000)
+        vRaw += -0x10000;
+    result.vBus = vRaw / 4096.0;
+}
+
+if (flags & 0x4) {
+    var iBoot = b[i];
+    i += 1;
+    result.boot = iBoot;
+}
+
+if (flags & 0x8) {
+    // we have temp, pressure, RH
+    var tRaw = (b[i] << 8) + b[i + 1];
+    if (tRaw & 0x8000)
+        tRaw = -0x10000 + tRaw;
+    i += 2;
+    var pRaw = (b[i] << 8) + b[i + 1];
+    i += 2;
+    var hRaw = b[i++];
+
+    result.t = tRaw / 256;
+    result.p = pRaw * 4 / 100.0;
+    result.rh = hRaw / 256 * 100;
+}
+
+if (flags & 0x10) {
+    // we have lux
+    var luxRaw = (b[i] << 8) + b[i + 1];
+    i += 2;
+    result.lux = luxRaw;
+}
+
+
+if (flags & 0x20)
+{
+	//Decode Rayco Sensor Data for Water Pressure & Water Level.
+	var ρ = 1000; 
+    var g = 9.81;
+    var wpRaw;
+	var wpFinal;
+    wpRaw  = (b[i] << 8) + b[i + 1];
+    i += 2;
+    //Convert pressure from hPa to kPa and store payload.
+    wpFinal = (wpRaw* 4 / 100.0) / 10;
+	result.wPressure = wpFinal;
+	result.wLevel = ((wpFinal)*1000) / (ρ*g);
+}
+
+// now update msg with the new payload and new .local field
+// the old msg.payload is overwritten.
+msg.payload = result;
+msg.local =
+    {
+        nodeType: "Catena 4450",
+        platformType: "Feather M0 LoRa",
+        radioType: "RF95",
+        applicationName: "Water-Level Sensor"
+    };
+
+return msg;
